+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix missing stack trace on UnobservedTaskException ([#2067](https://github.com/getsentry/sentry-dotnet/pull/2067))
 - Fix warning caused by db connection span closed prematurely ([#2068](https://github.com/getsentry/sentry-dotnet/pull/2068))
 - Attach db connections to child spans correctly ([#2071](https://github.com/getsentry/sentry-dotnet/pull/2071))
+- Improve MAUI event bindings ([#2089](https://github.com/getsentry/sentry-dotnet/pull/2089))
 
 ## 3.24.0